补充token滑动窗口及异步添入黑名单

+++ online-exam/src/main/java/com/exam/domain/Adminuser.java
@@ -1,7 +1,6 @@
 package com.exam.domain;
 
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
@@ -17,7 +16,7 @@ public class Adminuser {
      * ID号
      */
     @TableId(type = IdType.AUTO)
-    private Integer adminId;
+    private Long adminId;
 
     /**
      * 姓名
