+++ online-exam/src/main/java/com/exam/domain/Student.java
@@ -1,7 +1,6 @@
 package com.exam.domain;
 
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
@@ -17,7 +16,7 @@ public class Student {
      * ID
      */
     @TableId(type = IdType.AUTO)
-    private Integer studentId;
+    private Long studentId;
 
     /**
      * 姓名
