+++ online-exam/src/main/java/com/exam/domain/Teacher.java
@@ -1,7 +1,6 @@
 package com.exam.domain;
 
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
@@ -17,7 +16,7 @@ public class Teacher {
      * ID
      */
     @TableId(type = IdType.AUTO)
-    private Integer teacherId;
+    private Long teacherId;
 
     /**
      * 姓名
